refactor(auth)
