+++ app/src/auth/controller/auth.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+import { AuthService } from '../service';
 
 const registerUser = jest.fn();
 const generateToken = jest.fn();
