+++ app/src/auth/controller/auth.controller.ts
@@ -1,10 +1,10 @@
 import { Body, Controller, HttpCode, Post, Request, UseGuards } from '@nestjs/common';
-import { AuthRegistrationDto } from './dto/auth-registration.dto';
-import { AuthService } from './auth.service';
-import { UserWithoutPassword } from './authed-user';
-import { AuthLocalGuard } from './guard/auth-local.guard';
-import { AuthRefreshDto } from './dto/auth-refresh.dto';
-import { AuthJwtGuard } from './guard/auth-jwt.guard';
+import { AuthRegistrationDto } from '../dto/auth-registration.dto';
+import { AuthService } from '../service';
+import { UserWithoutPassword } from '../types';
+import { AuthLocalGuard } from '../guard/auth-local.guard';
+import { AuthRefreshDto } from '../dto/auth-refresh.dto';
+import { AuthJwtGuard } from '../guard/auth-jwt.guard';
 
 @Controller('/api/auth')
 export class AuthController {
