+++ app/src/auth/strategy/auth-local.strategy.ts
@@ -1,8 +1,8 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
-import { AuthService } from '../auth.service';
-import { UserWithoutPassword } from '../authed-user';
+import { AuthService } from '../service';
+import { UserWithoutPassword } from '../types';
 
 @Injectable()
 class AuthLocalStrategy extends PassportStrategy(Strategy) {
