+++ app/src/auth/types/authed-user.ts
@@ -1,3 +1,3 @@
-import { User } from '../user/entities/user.entity';
+import { User } from '../../user/entities/user.entity';
 
 export type UserWithoutPassword = Omit<User, 'password'>;
