+++ app/src/auth/types/index.ts
@@ -1,0 +1,2 @@
+export * from './authed-user';
+export * from './auth-jwt-payload.interface';
